ada: Simplify traversal procedures into traversal functions
Instead of using the generic routine Traverse_Proc to set a global flag when a
particular node is found, we can use its underlying routine Traverse_Func and
check if traversal has been abandoned. We already used this pattern in a number
of places; this patch merely applies it the remaining uses of Traverse_Proc.

Code cleanup; semantics is unaffected.

gcc/ada/ChangeLog:

	* exp_ch3.adb (Search_Access_Discriminant, Search_Current_Instance,
	Search_Internal_Call): Use traversal function instead of traversal
	procedure and remove associated global variables.
	* exp_util.adb (Search_Calls): Likewise.
	* sem_prag.adb (Contains_Loop_Entry): Likewise.
	* sem_util.adb (Mentions_Post_State): Likewise.